requirements.txt
fastapi==0.116.1
click==8.1.7
tortoise-orm==0.25.1
asyncpg==0.30.0
aiomysql==0.2.0
celery==5.5.3
redis==6.2.0
bcrypt==4.3.0
passlib==1.7.4
python-jose==3.5.0
python-dotenv==1.1.1
uvicorn==0.35.0
uvloop==0.21.0
pydantic[email]

elasticsearch==9.1.0
openai==1.98.0
llama-index==0.13.0
#llama-index-readers-web==0.5.0
llama-index-readers-web[trafilatura]==0.5.0
aiofiles==24.1.0
PyPDF2==3.0.1
PyMuPDF==1.26.3
pdfplumber==0.11.7
pytesseract==0.3.13
pillow==11.3.0
Brotli==1.1.0
trafilatura==2.0.0
